+++ Package.swift
@@ -1,0 +1,19 @@
+// swift-tools-version:5.3
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+import PackageDescription
+
+let package = Package(
+    name: "KlarnaMobileSDK",
+    products: [
+        .library(
+            name: "KlarnaMobileSDK",
+            targets: ["KlarnaMobileSDK"]),
+    ],
+    dependencies: [],
+    targets: [
+        .binaryTarget(
+            name: "KlarnaMobileSDK",
+            path: "ios/XCFramework/basic/universal/KlarnaMobileSDK.xcframework"
+        )
+    ]
+)