+++ ios/xcode-10.3-slim/KlarnaMobileSDK.framework/Headers/KlarnaMobileSDK-Swift.h
@@ -193,6 +193,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, closed) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -264,6 +277,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -337,6 +352,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -431,6 +451,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, closed) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, closed) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, closed) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -438,8 +476,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -641,6 +679,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -671,8 +714,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+
 
 
 
@@ -875,6 +926,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, closed) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -946,6 +1010,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -1019,6 +1085,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -1113,6 +1184,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, closed) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, closed) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, closed) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -1120,8 +1209,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -1323,6 +1412,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -1353,8 +1447,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+