+++ ios/xcode-11.7-fat/KlarnaMobileSDK.framework/Headers/KlarnaMobileSDK-Swift.h
@@ -217,6 +217,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, open) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -288,6 +301,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -361,6 +376,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -455,6 +475,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, open) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, open) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, open) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -462,8 +500,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -665,6 +703,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -695,8 +738,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+
 
 
 
@@ -919,6 +970,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, open) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -990,6 +1054,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -1063,6 +1129,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -1157,6 +1228,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, open) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, open) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, open) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -1164,8 +1253,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -1367,6 +1456,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -1397,8 +1491,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+
 
 
 
@@ -1625,6 +1727,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, open) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -1696,6 +1811,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -1769,6 +1886,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -1863,6 +1985,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, open) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, open) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, open) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -1870,8 +2010,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -2073,6 +2213,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -2103,8 +2248,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+
 
 
 
@@ -2327,6 +2480,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, open) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -2398,6 +2564,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -2471,6 +2639,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -2565,6 +2738,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, open) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, open) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, open) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -2572,8 +2763,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -2775,6 +2966,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -2805,8 +3001,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+