+++ ios/xcode-11.7-slim/KlarnaMobileSDK.framework/Headers/KlarnaMobileSDK-Swift.h
@@ -213,6 +213,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, open) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -284,6 +297,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -357,6 +372,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -451,6 +471,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, open) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, open) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, open) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -458,8 +496,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -661,6 +699,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -691,8 +734,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+
 
 
 
@@ -915,6 +966,19 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK16KlarnaDebugEvent")
+@interface KlarnaDebugEvent : NSObject
+- (nonnull instancetype)init SWIFT_UNAVAILABLE;
++ (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
+@end
+
+typedef SWIFT_ENUM(NSInteger, KlarnaDebugEventType, open) {
+  KlarnaDebugEventTypeLog = 0,
+  KlarnaDebugEventTypeNetwork = 1,
+  KlarnaDebugEventTypeMessage = 2,
+};
+
+
 SWIFT_CLASS("_TtC15KlarnaMobileSDK11KlarnaEvent")
 @interface KlarnaEvent : NSObject
 @property (nonatomic, readonly, copy) NSString * _Nullable bodyString;
@@ -986,6 +1050,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 @protocol KlarnaHybridSDKEventListener;
 
 @interface KlarnaHybridSDK (SWIFT_EXTENSION(KlarnaMobileSDK))
+/// Arbitrary data as string that will be sent to Klarna components in handshake with the SDK.
+@property (nonatomic, copy) NSString * _Nullable klarnaInitData;
 /// Initialize the Klarna Mobile SDK in hybrid mode.
 /// warning:
 /// This initializer is deprecated. Use the new one and just initialize the SDK and
@@ -1059,6 +1125,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK15KlarnaHybridSDK")
 /// \param callback Closure to handle message events sent to merchants.
 ///
 - (void)registerEventListenerWithCallback:(void (^ _Nonnull)(KlarnaEvent * _Nonnull))callback;
+@end
+
+
+SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaHybridSDKDebug")
+@interface KlarnaHybridSDKDebug : KlarnaHybridSDK
 @end
 
 
@@ -1153,6 +1224,24 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 + (nonnull instancetype)new SWIFT_UNAVAILABLE_MSG("-init is unavailable");
 @end
 
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMEnvironment, open) {
+  KlarnaOSMEnvironmentDemo = 0,
+  KlarnaOSMEnvironmentProduction = 1,
+  KlarnaOSMEnvironmentPlayground = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMRegion, open) {
+  KlarnaOSMRegionEu = 0,
+  KlarnaOSMRegionNa = 1,
+  KlarnaOSMRegionOc = 2,
+};
+
+typedef SWIFT_ENUM(NSInteger, KlarnaOSMTheme, open) {
+  KlarnaOSMThemeDark = 0,
+  KlarnaOSMThemeLight = 1,
+  KlarnaOSMThemeAutomatic = 2,
+};
+
 @class NSCoder;
 
 /// OSM Placement View
@@ -1160,8 +1249,8 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK20KlarnaMobileSDKError")
 /// <code>render()</code> is called.
 /// The view self-sizes vertically. Add to parent with <code>addSubview()</code> and set outer
 /// constraints (top, left, leading, trailing).
-SWIFT_CLASS("_TtC15KlarnaMobileSDK18KlarnaOSMView_Beta")
-@interface KlarnaOSMView_Beta : UIView
+SWIFT_CLASS("_TtC15KlarnaMobileSDK13KlarnaOSMView")
+@interface KlarnaOSMView : UIView
 - (nonnull instancetype)initWithFrame:(CGRect)frame SWIFT_UNAVAILABLE;
 - (nullable instancetype)initWithCoder:(NSCoder * _Nonnull)coder SWIFT_UNAVAILABLE;
 @end
@@ -1363,6 +1452,11 @@ SWIFT_CLASS("_TtC15KlarnaMobileSDK17KlarnaPaymentView")
 @end
 
 
+SWIFT_CLASS("_TtC15KlarnaMobileSDK22KlarnaPaymentViewDebug")
+@interface KlarnaPaymentViewDebug : KlarnaPaymentView
+@end
+
+
 /// General class that envelops WKWebView
 SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 @protocol KlarnaWebView
@@ -1393,8 +1487,16 @@ SWIFT_PROTOCOL("_TtP15KlarnaMobileSDK13KlarnaWebView_")
 
 
 
+
+
+
+
+
+
 @interface WKWebView (SWIFT_EXTENSION(KlarnaMobileSDK)) <KlarnaWebView>
 @end
+
+